+++ viadot/sources/__init__.py
@@ -23,6 +23,8 @@ from .sqlite import SQLite
 from .mindful import Mindful
 from .customer_gauge import CustomerGauge
 
+from .vid_club import VidClub
+
 # APIS
 from .uk_carbon_intensity import UKCarbonIntensity
 from .eurostat import Eurostat